Adds callback on double signed message is creation
Dispersy.create_signature_request() now has success_func and
success_args parameters.  The success_func is called when a double
signed message is successfully created.

This can be used to immediately handle or process this message outside
of the standard callback that occurs when any other message is received.
For example, when you want to immediately forward this new message to
the other peer who participated in signing.